Update README.md
About the Movie Pitch Perfect is a musical comedy film about college a cappella groups competing to win the national championship. It focuses on Beca Mitchell, a college student who joins the all-girls group The Barden Bellas and helps transform their sound with her creativity and love for remixing music.